module/tests/requirements.txt
pytest
pytest-cov
black
coverage
kubernetes
cryptography
pyyaml